Moved deferreds and core event handlers into connection
